Rename uw_record -> auditor
* access via underworld3.auditor
* test included in test_0005_utils.py
